+++ m3cli/plugins/apply-terraform-template.py
@@ -9,24 +9,21 @@ def create_custom_request(request):
     request.parameters['task'] = 'PLAN'
     if request.parameters.get('variables'):
         variables = request.parameters.pop('variables')
-        value = list(variables.values())[0]
-        if ',' in value:
-            type = 'LIST'
-            value = value.replace(' ', '').strip('][').split(',')
-        elif '=' in value:
-            type = 'MAP'
-            temp = value.replace(' ', '').split('=')
-            value = {temp[0]: temp[-1]}
-        else:
-            type = 'STRING'
-        name = list(variables.keys())[0]
-        request.parameters['variables'] = {
-            name: {
+        request.parameters['variables'] = {}
+        for name, value in variables.items():
+            if ',' in value:
+                type = 'LIST'
+                value = value.replace(' ', '').strip('][').split(',')
+            elif '=' in value:
+                type = 'MAP'
+                temp = value.replace(' ', '').split('=')
+                value = {temp[0]: temp[-1]}
+            else:
+                type = 'STRING'
+            request.parameters['variables'][name] = {
                 'type': type,
                 'value': value,
                 'sensitive': True,
                 'name': name
             }
-        }
     return request
-