+++ tests/test_plugins/test_delete_storage.py
@@ -6,30 +6,16 @@ class TestDeleteStorage(TestPluginsBase):
 
     def test_delete_storage_success_request(self):
         self.request.parameters = {
-            'region': 'AZURE',
-            'resourceGroup': 'some_group'
+            'region': 'AZURE'
         }
 
         expected_result = {
             'region': 'AZURE',
-            'resourceGroup': 'some_group',
-            'params': {
-                'resourceGroupName': 'some_group'
-            }
+            'params': {}
         }
 
         result = self.execute_create_custom_request()
         self.assertEqual(result.parameters, expected_result)
-
-    def test_delete_storage_resource_group_missed_request(self):
-        self.request.parameters = {
-            'region': 'AZURE'
-        }
-
-        with self.assertRaises(AssertionError) as context:
-            self.execute_create_custom_request()
-        self.assertEqual(str(context.exception), 'Parameter resource-group'
-                                                 ' is required for AZURE cloud')
 
     def test_delete_storage_availability_zone_missed_request(self):
         self.request.parameters = {